Fix compilation error in DelaunayTriangulationTest
This fixes a compilation error in the
DelaunayTriangulationTest class in which
qCompare<unsigned int, unsigned long> was not found.

+++ tests/auto/chemkit/delaunaytriangulation/delaunaytriangulationtest.cpp
@@ -74,7 +74,7 @@ void DelaunayTriangulationTest::joe89()
     expectedTetrahedra.append(QVector<int>() << 4 << 5 << 6 << 7);
 
     std::vector<std::vector<int> > tetrahedra = triangulation.tetrahedra();
-    QCOMPARE(tetrahedra.size(), static_cast<unsigned long>(expectedTetrahedra.size()));
+    QCOMPARE(tetrahedra.size(), size_t(expectedTetrahedra.size()));
 
     for(unsigned int i = 0; i < tetrahedra.size(); i++){
         std::vector<int> tetrahedron = tetrahedra[i];
